+++ src/PortalBase/StorageKey/MappingNodeKeyCollection.php
@@ -21,7 +21,7 @@ class MappingNodeKeyCollection extends AbstractCollection
         );
     }
 
-    public function unique(): self
+    public function asUnique(): self
     {
         return $this->uniqueByContains();
     }