+++ src/StorageBase/FileReferenceRequestKeyCollection.php
@@ -21,7 +21,7 @@ final class FileReferenceRequestKeyCollection extends AbstractCollection
         );
     }
 
-    public function unique(): self
+    public function asUnique(): self
     {
         return $this->uniqueByContains();
     }