+++ src/StorageBase/JobKeyCollection.php
@@ -21,7 +21,7 @@ class JobKeyCollection extends AbstractCollection
         );
     }
 
-    public function unique(): self
+    public function asUnique(): self
     {
         return $this->uniqueByContains();
     }