+++ src/StorageBase/MappingNodeStructCollection.php
@@ -20,7 +20,7 @@ class MappingNodeStructCollection extends AbstractObjectCollection
         );
     }
 
-    public function unique(): self
+    public function asUnique(): self
     {
         return $this->uniqueByContains();
     }