+++ src/StorageBase/RouteKeyCollection.php
@@ -21,7 +21,7 @@ class RouteKeyCollection extends AbstractCollection
         );
     }
 
-    public function unique(): self
+    public function asUnique(): self
     {
         return $this->uniqueByContains();
     }